mcp-server-telegram/pyproject.toml
[project]
name = "mcp-server-telegram"
version = "0.1.0"
description = "MCP Server for sending messages to Telegram channels"
authors = [
    { name = "XyAgent", email = "agent@xyber.inc" },
]
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "mcp>=1.8.0,<2.0.0",
    "pydantic-settings>=2.0",  
    "uvicorn[standard]",
    "fastmcp>=2.3.4",
    "fastapi>=0.115.12",
    # Telegram-specific dependencies
    "python-telegram-bot>=20.0",
]

[dependency-groups]
dev = [
    "pytest",
    "httpx", 
    "ruff",
    "black",
    "isort",
]
debug = [
    "debugpy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_server_telegram"]

# --- Tool Configurations --- #

[tool.ruff]
line-length = 88
target-version = "py312"
select = ["E", "F", "I", "W", "C90", "N", "D"]
ignore = ["D100", "D104", "D107", "D203", "D212"]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 88